renamed app.routing to be consistent with angular docs and our other modules
